fixed getNearbyParks
